requirements.txt
accelerate==1.12.0
anaconda-anon-usage==0.7.4
anaconda-auth==0.10.0
anaconda-cli-base==0.6.0
annotated-types==0.6.0
anyio==4.12.1
archspec==0.2.5
boltons==25.0.0
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
conda==25.9.1
conda-anaconda-telemetry==0.3.0
conda-anaconda-tos==0.2.2
conda-content-trust==0.2.0
conda-libmamba-solver==25.4.0
conda-package-handling==2.4.0
conda_package_streaming==0.12.0
cryptography==46.0.3
distro==1.9.0
filelock==3.20.1
frozendict==2.4.6
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jaraco.classes==3.4.0
jaraco.context==0.0.0
jaraco.functools==4.1.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
libmambapy==2.3.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
menuinst==2.4.1
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.21.0
packaging==25.0
pip==25.2
pkce==1.0.3
platformdirs==4.5.0
pluggy==1.5.0
psutil==7.1.3
pyairports==0.0.1
pycosat==0.6.6
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.10.1
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.3
readchar==4.2.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.7.0
SecretStorage==3.4.0
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
truststore==0.10.1
typer==0.17.4
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
wheel==0.45.1
zstandard==0.24.0
